Add instrumention pass for indirect function calls
This commit adds a pass that inserts instrumentation for
profiling the targets of indirect function calls. The pass
currently inserts annotations that record an identifier for
each function. This identifier (called a function number) is
used by the profiling runtime to identify the function that
is called through an indirect branch.

We still need to insert the callbacks into the profiling
runtime that will record the address of the indirect
function being called.