Added JIT listener for indirect function call profiling
This commit adds a JITEventListener that is used for the
indirect branch profiling. The listener examines the
function that was emitted looking for the function number
annotation.

We still need to add the callback to the profiling runtime
that will record the association between the function number
and the function address so that it can be used to profile
the targets of indirect branches.