Add runtime for profiling indirect function calls
This implement the runtime callbacks needed to profile the
targets of indirect function calls. These callbacks are used
by the instrumentation pass to record the actual target of
an indirect function call and by the JIT to build a mapping
from code addresses to the function residing at that
address.