Added profiling callback in IFC JIT listener
This commit adds a callback to the profiling runtime to the
JIT listener for Indirect Function Call profiling. The
listener waits for the ResolvedLazyStub event and then
issues a callback into the profiling runtime passing the
stub address and its corresponding function number so that
the profiling runtime can associate that address to the
correct function.