Update IFC profiling runtime to track target addr
This commit updates the Indirect Function Call profiling
runtime to keep the profiling targets by machine address
rather than function number. A separate map is kept that
maps addresses to function numbers. Keeping the profiling
by address allows us to not have to have a function number
entry for address before the first time we see the address.
Now we lazily associate the address to a function number in
the profiling results when the profiling is finished.