Ouput IFC profile data in compact binary format
This commit adds the ability to dump the indirect function
call profiling data in a compact binary format. The
profiling data matches the general form of other profiling
data with a header that records the number of entries of
profiling data and an identifier of the profiling type:

   <-- 4 bytes --> <-- 4 bytes -->
   +--------------+---------------+
   | IFCInfo      | # of Records  |
   +--------------+---------------+

and then a number of entries that contain records of the
form:

   <-- 4 bytes --> <-- 4 bytes  --> <--   8 bytes  -->
   +---------------+----------------+------------------+
   |   CallSite#   |Target Function#|     Count        |
   +---------------+----------------+------------------+

These recrds will be read by later passes that want to make
use of the profiling data.