Implement IndirectFunctionCallProfileLoader pass
This pass is used to read indirect function call profile
data from files generated by the IFC profiling runtime. The
loader will read the profiling data file and populate an
IndirectFunctionCallProfileInfo object with the data from
the file.

The data can then be accessed from other passes by using a
  getAnalysisUsage<IndirectFunctionCallProfileInfo>();
call to get a the profile data and adding -load-ifc-profile
option to the opt command line.