Create library for IFC profiling support routines
This commit creates a new Profile library that holds
routines used for indirect function call profiling. The
routines are used to compute function and call site
numbering to ensure that the numbering used for
instrumentation will match the numbering used when reading
profile data files.

Additionally, we have updated the instrumentation and
profile reading passes to use the new library support
routines.