Add pass to build traces based on profile data
This commit adds a -build-traces pass that can be used to
construct program traces based on profiling data. The trace
profile is read and then used to build new program traces.
The traces are built by cloning the functions in the trace
and modifying the blocks on the trace so that they call the
newly cloned functions.