Allow traces to include calls to intrinsic functions
This commit adds an option to the trace profiler that will
not count calls to intrisic functions as a break in the
trace. If this option is set then we don't insert a call to
the trace runtime to let it know that an external function
call (the intrinsic) is about to execute.

We also update the trace builder to allow blocks with
muliple calls as long as all calls except one is a call to
an intrinsic.

We want to allow intrinsic calls in the trace so that we can
have traces that include a call to libm functions such as
pow which show up as intrinsics. We are assuming the the
intrisic functions will always return to the caller so from
the llvm point-of-view the trace is still continuous.