Profile trace exit percents
This commit adds the ability to profile trace exits. After a trace
is recorded we will check each time we enter the trace if we make
it all the way through. If we don't we can record which block
caused us to leave the trace.

This commit updates the trace profiling to record the exit info
and read it from the trace profiles. The llvm-prof tool now
displays exit percents next to each block.